Changes to announcements, mostly for AI Auto Play
During AI Auto Play, the color of announcements and their map locations are no longer discarded; i.e. one can now inspect the location of an announcement by clicking on it in the Event Log once Auto Play has ended. (However, even in Debug mode, this only works if the location is revealed to the active player.) [advc.127]

Added location info to numerous announcements so that the camera moves somewhere when the announcement is clicked on in the Event Log. E.g. clicking on a declaration of war now moves the camera to the capital of the civ that declared war (if that capital is revealed). Added a loading screen hint about clicking on stuff in the Event Log. [advc.127b]

Canceled Defensive Pacts are announced and recorded in replays. [advc.106f]

Discovered tech demoted to a minor event; no longer kept in the Event Log indefinitely. [advc.106b]

AI ships don't blockade Barbarian cities (or at least not as frequently as in BtS). [advc.033]

Fixed a bug in advc.130r that had caused "You granted us our independence" to be shown twice in help text.