War Trades alert no longer triggers when a war trade becomes unavailable. [advc.210a]
Fixed an issue with alerts getting triggered already during AI turns (and thus sometimes being out of date when displayed) when the human player is not in player slot 0. [advc.210]

Fixed a bug in UWAI that had made it too easy to hire an AI civ for war when it is about to attack the human player.

AI cancels its war preparations upon forming a Permanent Alliance with a human. [advc.001]