Went through the 2nd half of the AdvCiv changelog in the manual and addressed some open issues (DLL included)
Extended change [advc.001c] (GP birth probabilities in GP help text computed based on current GPP per turn) so that those probabilities are also shown directly on the GP bar and on the BUG GP progress bar (which is still disabled by default).

Change [advc.120c] (espionage slider hidden when at 0) can be disabled on the "General" tab of the BUG menu.

Got change [advc.001h] to work: The full city radius is highlighted when a Settler is selected and no yield icons are shown. The latter change is implemented through a bit of a hack: I've replaced CvUnit::isFound with a new function CvUnit::canFound, so that the EXE can't tell anymore whether a Settler is selected.

Tweaked [advc.120]: AI more likely to train Spies during war preparations (as in K-Mod), but not once war is declared.

Merged previously omitted parts of this K-Mod 1.45 change:
"The AI now counts feature defensive modifiers as having negative value
when adjacent to cities; and so they are more likely to chop forests
that are next to cities."
https://github.com/karadoc/Civ4-K-Mod/commit/f3819966a65ce8e04747aaf4d562a5fd18a0e5fe
However, this is only relevant if change [advc.012] is reverted; that change sets feature defense to 0 for Forest and Jungle and introduces a new XML tag "RivalDefense".

Changes [advc.121] (AI placement of Forts) and [advc.314] (promotions from goody huts partly based on surrounding terrain features) now work correctly with the RivalDefense ability of [advc.012].

Addition to [advc.250b]: No free wins against Barbarians if the R&F option is enabled.

[advc.300] no longer shows on the Settings tab on which turn Raging Barbarians can first appear because that info isn't really needed.

Barbarian Spearmen no longer stop appearing when the Barbarians reach the Medieval era. [advc.301]

Disabled the part of [advc.307] that had allowed Barbarian cities with local Ivory to train War Elephants.

Barbarian city placement exempted from changes to AI found value in the previous commit. [advc.108]
And some other minor changes to Barbarian found value. [advc.303]

When deciding whether to move the starting Settler, the AI assumes a low value for unexplored tiles. (Normally, 0 value is assumed for unexplored tiles.) [advc.108]

R&F: AI refuses to talk if its war utility is 20 or higher (threshold was 10 previously). [advc.705]