AI starting units placed separately; cosmetic changes from Unofficial Patch (no DLL/ manual committed)
Exploration units that civs receive at game start are placed on a different tile than the starting Settler and defensive units. This was already the case for human civs, so the change makes the starting positions of human and AI civs more alike. The extra visibility makes it easier for the AI to move its starting Settler (although this was already working pretty well after the previous commit). [advc.108]

Merged a few cosmetic changes from Roamty's Unofficial Patch 1.7 [advc.002]
https://forums.civfanatics.com/threads/unofficial-3-19-patch-1-7.637808/