A few additions to the previous commit:
I noticed that the BUG help file said that the GG bar "will not automatically update after a battle, but you must enter a screen and exit out for it to display properly.” Fixed this by setting a dirty-bit in CvPlayer::setCombatExperience. [advc.004]

Some (final) revisions of the BUG menu help texts [advc.004]

Removed the xmlKey parameter from the getBUGOption... functions because the callee didn't use it and the (actual) arguments were cluttering up the code. [advc.003]
And I've aligned the iDefault arguments (that apply only if the option isn't found) with my changes to the default values in BUG's XML config.