AI contacts human players to sue for peace or capitulate
Finally got [advc.134a] to work. It's a workaround for the bug in the EXE that prevents AI-to-human peace offers from being shown. CvTeam::isAtWar now feigns peace at just the right moment to stop the EXE from discarding the peace offer. The downside is that, to ensure the proper order of isAtWar calls, CvPlayer::canContact and CvPlayer::canTradeItem now have to call a different function (CvTeam::isAtWarInternal, with ::atWar in CvGameCoreUtils as an alias) to check at-war status. And, in terms of the challenge posed by the AI, the peace offers (which have to include a discount to be worth considering) can only hurt the AI against a competent player.

I've made a couple of adjustments to the timing of AI peace offers and the items that are demanded or offered. In particular, I changed the ContactRand values (was 1 chance in 20 for all leaders) and ContactDelay (was 10 turns for all leaders) of some AI leaders:
Delay decreased to 5: Gandhi
Delay increased to 15: Boudica, Sitting Bull, Tokugawa
Rand decreased to 10: Gandhi
Rand decreased to 12: Mansa Musa, Pacal, Joao, Hatshepsut, Ashoka, Huayna Capac
Rand decreased to 15 for 17 of the 52 leaders.
These timing values only affect AI-to-human peace if UWAI is enabled; otherwise (legacy AI), ContactRand also affects peace deals between AI civs.

This commit will be difficult to test properly because AI Auto Play doesn't help with AI-to-human offers.

I've moved some K-Mod code for creating AI proposals around to avoid code duplication; hopefully no functional difference. [advc.003]

Fixed a bug in [advc.001e] (cancellation of outdated AI defensive pact offers and stop-trading demands): wrong order of remove and delete calls.

In Debug mode, holding down [Alt] while hovering over a leader portrait shows the leader's id (player slot number). [advc.007]

Plus a little bit of random refactoring in CvCityAI.cpp and CvGame.cpp.