Bugfixes (in AdvCiv code); enhanced current deals on the Trade screen
Fixed a bug introduced in a commit 2 weeks ago (i.e. it's in the beta version 0.94e) that sets the UWAI threshold for war trades to -5 (instead of -33). This had made the AI almost always unwilling to be hired for war.

Fixed a minor bug (also introduced in some recent commit) that had triggered a spurious willing-to-talk alert after hiring an AI civ for war. [advc.104i]

Fixed a minor (potential) issue with [advc.046] (unloading of units in order to pick up stranded units).

Found a way to obtain the CvDeal object that corresponds to the trade item that the EXE provides when calling CvDLLWidgetData::parseTradeItem and CvPlayer::getItemTradeString. This allowed me to make the following improvements to the list of current trades on the Trade screen: [advc.072]
* Hover text for trade items shows the deal that the item belongs to (e.g. "Banana to Peter for Rice and 2 gold per turn" when hovering over "Banana") and whether/ when that deal can be canceled.
* No trade denial is shown when hovering over resources that are being traded.
* The "Deal Turns Left" option on the "Advisors" tab of the BUG menu now has four choices: Show the number of remaining turns
- only on the "Active" tab of the Foreign Advisor screen,
- only in help text (still the default), and new:
- only in the list of current trades on the Trade screen, or
- both on the "Active" tab and the Trade screen.
The new choices will e.g. show "Open Borders (8 Turns)" in both columns of the Trade screen when Open Borders have been signed two turns ago.

Double-clicking the city tile no longer exits the city screen. (Because this shortcut collides with the one for enabling citizen automation - a single click on the city tile.) [advc.004t]

Added the single click on the city tile to the list of shortcuts in Sevopedia. [advc.004t]

Loading-screen hint added about assigning a working city to tiles that multiple cities can work. [advc.008d]

When an AI civ checks if there are any good resources near a potential city site, resources in tiles that the civ already owns are counted so long as they aren't within the radius of any existing city. Such resources had previously been disregarded (which had prevented the AI from settling Crete in my EuroWorld scenario). [advc.031]

Refactoring in CvPlayerAI::AI_updateFoundValues and a little bit in CvPlayerAI::AI_foundValue and CvPlayer::setTurnActive. [advc.003]