Changes to AI greetings
Only about half of the AI leaders (based on CONTACT_DEMAND_TRIBUTE) brag about their best units. When a braggart AI has a new best unit, it brags about that unit at the first opportunity (unless the best unit is known anyway because AI technologies and resources are visible to the human player) and, if contacted again, with a random chance as in BtS. This eliminates the need for contacting the AI multiple times to find out its best unit in the early game. [advc.079]

Partially revised the CONTACT_DEMAND_TRIBUTE divisors in the LeaderHead XML file. For example, Shaka, Ragnar, Mehmed were among the least likely AI leaders to demand tribute (divisor 1000); now they're among the most likely (divisor 25), whereas e.g. De Gaulle and Isabella are no longer very likely to make demands. I've done this to support the change to bragging (want the early warmongers to brag), but also to match the historical background better. [advc.005a]

The AI remembers when it has warned a human player against trading with a particular enemy, and is more likely to warn if it hasn't warned before and less likely otherwise. Mainly to reduce the frequency of the accompanying leader animations, which are somewhat jarring.
After the first era, the AI says "we would encourage you to stop trading ..." in the greeting only if the human player has annual deals with the worst enemy. [advc.079]

And some unrelated changes:

Warning about XP loss added to unit upgrade button. [advc.080]
Related CFC thread: https://forums.civfanatics.com/threads/k-mod-far-beyond-the-sword.407049/page-149#post-15396176

Fixed an issue with [advc.004h] (city radius colored around selected Settler): The color had lingered after deleting a Settler, and occasionally, after founding a city.

In scenarios, the start era is set to the mean of the player eras. This blocks some problematic Ancient wonders in the Earth1000AD scenario. See this CFC post: https://forums.civfanatics.com/threads/earth-1000ad-shadow-game-sunset-invasion-invading-europe-with-aztec.643333/#post-15396111