Change overflow handling (work in progress)
[advc.064b]
On-screen message added that tells the player when a city has used excess overflow to train additional units.

Set the upper bound for the overflow that a city can store to the maximum of the city's food store size and (as before) its base production rate. The cost of the last completed production order no longer matters. This should fix balance problems with AI cities training multiple Archers per turn through Slavery on the highest difficulty levels.

However, I'm going to disable and remove the K-Mod mechanism for training multiple units per turn, so this commit is only an interim result.