Some misc. tweaks, mostly to AdvCiv changes
Increased the AI production weight a bit when the AI is not in Slavery. [advc.110]

Increased AI trade value for strategic resources that boost buildings currently under construction. For one thing, this should make sure that the AI doesn't trade away its only Stone or Marble source while building a wonder requiring that resource. [advc.036]

AI won't ask for tribute or a gift during a peace treaty except, in the case of a gift, when the AI remembers having granted a gift request. [advc.144]

When units in a transport are woken up by [advc.075], the units are now put in a single selection group for convenience.

Fixed a small issue with [dlph.26] (changes the order of triggered wars): When declaring war on an embargoed civ, the refuse-to-talk alert had misreported that civ as willing to talk.

UWAI: The cache for preventing war utility from getting recomputed during UI interactions is now invalidated in CvGame::update. This should fix issues with the willing-to-talk indicator on the scoreboard being out of date. [advc.104l]

Meant to make a small amendment to [advc.130p] (AI may refuse to sign Open Borders on account of having "nothing to gain"), then changed my mind and left my changes in the code as comments.

The choose-production popup always names the city (in BtS, it says "in this city" sometimes). [advc.004]

The "older messages" marker is inserted into the Event Log even if the AdvCiv option for opening the log automatically is disabled. [advc.106b]

Added a function for casting CvTeam pointers to CvTeamAI references, just like the ones added earlier for CvPlayer, CvCity, CvUnit and CvGame. Replaced occurrences of 'GET_PLAYER(getID())' in CvPlayer.cpp and 'GET_TEAM(getID())' in CvTeam.cpp. [advc.003]

Removed numerous unnecessary getInstance calls from CvGlobals.cpp.