A few changes concerning AI civic requests and naval invasions
Changed Suryavarman's favorite civic from Organized Religion to Caste System because asking others to change into Organized Religion (or Theocracy, Pacifism) generally doesn't make much sense unless the state religions happen to match. This was especially annoying in Sury's case because his change-civics requests apply a -2 relations penalty when rebuked. [advc.005a]
(An earlier version of AdvCiv had already changed his favorite religion from Buddhism to Hinduism for historical accuracy.)

When a player accepts an AI request to adopt a civic or religion and then switches out of that civic or religion, the relations bonus is forgotten faster (4 times faster to be exact). Likewise, the relations penalty from refusing to switch is forgotten faster (3.5 times) when the player makes the requested switch later on. Previously, AdvCiv had suspended the relations modifier in such situations, but I've come to think that this makes it too unattractive to accept AI requests. [advc.145]

Disabled [advc.162] (units that enter enemy borders on the turn that war was declared lose all remaining movement points). I don't think it's quite worth the added complexity all things considered, but the change works alright and can be re-enabled through GlobalDefines_advc.xml.

When an AI stack is about to declare war, it tries to avoid moves through the enemy's territory that would lead to a declaration of war earlier than necessary. There was already BBAI/K-Mod code for this, but that only stopped the AI from ending a move in the enemy's territory; passing through in a single move wasn't covered. [advc.082]

Another small tweak to [advc.048] to prevent the "Press ALT for best odds" hint from being shown unnecessarily when units are tied for best odds.

A few more tech cost changes. [advc.910,advc.251]

When a game ends, the "Demographics" tab shows information as if the player knew the demographics of all rivals. [advc.077]
(This wasn't an issue in BtS because the "Demographics" tab never hid any information.)

For autosaves in multiplayer, it is now assumed that the game name "chipotle" implies that this is a test game run on a single machine. Only players with even id numbers create autosaves then (to avoid error popups from writing to the same file at the same time). [advc.135c]