Minor changes to unit plot list and AI attitude cache
Bug fixed in the PLE component: The names of unit owners had been shown in random colors. [advc.069]

No indicator circles are shown in the unit plot list for foreign units because it's always a white circle (meaning that the unit has full moves but isn't expecting orders). This change only applies when playing with the BtS plot list (BUG and PLE drawing methods disabled on the "Unit Icons" tab of the BUG menu). [advc.069]

Fixed an out-of-date issue with the AI attitude cache and AdvCiv's "you're getting ahead of us" modifier. [advc.130c]

When showing a breakdown of relations modifiers, the UI checks if the sum of the modifiers matches the cached total relations value. If not, an assertion message is shown and the attitude cache is updated. (Unless it's a networked multiplayer game.) [advc.sha]