More scoreboard changes; expanding scoreboard enabled by default.
Enabled the expanding scoreboard by default after some testing and some more changes: [advc.085]
* Expanded scoreboard no longer collapses when the cursor is moved onto a gap in the tabular layout.
* Hover text for the power ratio column added. Explains what the ratio means and why rival demographics are visible: by how many espionage points the threshold for seeing demographics or research (whichever applies) is exceeded. Power ratio column enabled by default, but only appears when the scoreboard is expanded.
* Golden Age/ Anarchy column added. Appears by default when the scoreboard is expanded. Hover text says how many turns remain.

In the current-research column of the scoreboard, show a progress percentage of 0 when no tech is currently researched. [advc.085]

In the power ratio column of the scoreboard, don't use a different color for high or low power ratios of allies. [advc.085]

Changed the default colors of power ratios so that 0-100% rival power is shown in white (low ratio), 101-130% in orange (normal ratio) and 131% or more in red (high ratio). [advc.004]

Added help text about [advc.124] (can only trade with revealed cities) to the list of circumstances that prevent foreign trade as shown by the BULL hover text for the trade network icon on the scoreboard ("Trade Hover").

Added a macro PLAYER_TEXT_COLOR to avoid having to pass the R, G, B and A values separately to SETCOLR. [advc.003]