AI adjustments for military victory strategies
The AI seemed to be much more successful with peaceful victory strategies. The following changes in the UWAI component are a step toward remedying that:

Reduced fairness cost for dogpile wars, especially in the late game.

Decreased the war utility counted for loathing and for kingmaking (the part that makes the AI try to thwart an enemy's victory).
As these (pretty minor) changes had resulted in far fewer early wars, I've increased the utility counted for greed.

Reverted an earlier change (from October) that had caused the AI to count war utility for coming close to a military victory already at stage 2 of a military victory strategy. I think this had mainly resulted in the AI being quicker to start wars, but, when a military victory is a long-term prospect, it's more important to build up military power. Instead of affecting war utility, the first two stages of the military victory strategies now have a greater impact on civics and research choices (replacing BBAI's AI_getWarThresholds function).

When estimating likely winners, civs at stage 2 of a peaceful victory on a small landmass are treated as if their game score was 33% greater than it actually is.

Some changes to AI_targetCityVal to make the AI target civs that aim for a peaceful victory earlier and more reliably. [advc.104d]

Mostly unrelated:
* Corrected a small oversight in the affection cost calculation.
* Made the AI a bit less willing to raze Barbarian cities. [advc.116]