XML option for Dawn-of-Man after map-regen added
Found a way to center the camera after regenerating the map without showing the Dawn-of-Man screen: Wait for the game to update a few times (CvGame::update). [advc.004j]

Since I was already using that trick for [advc.085] and [advc.001w], I've added a function CvGame::setUpdateTimer (akin to deferCall in BugUtil.py); that code is tagged with [advc.003r].

I still like it a bit better with the Dawn of Man screen because, this way, the player doesn't see the old map fade to black. But there's now an option SHOW_DAWN_AFTER_REGEN in GlobalDefines_advc.xml.