Renamed getX_INLINE, getY_INLINE to getX, getY
Through automated search and replace, replacing "_INLINE" with an empty string. Then manually changed some linebreaks (now that the function names are shorter, getX and getY should always be on the same line when both are passed as function arguments) and removed some stray whitespaces that I happened to come across.

Follow-up on this commit:
https://github.com/f1rpo/AdvCiv/commit/32d7dd438a3b00b05824be41ec01da28b1b4ffdd

Coming up: Rename the remaining "INLINE" functions.