v0.96 release candidate and partly revert previous commit
Manual and DLL (final release build) updated.

AI_isWarPossible was correct after all -- the "Always War" option means that the human players are at war with the AI civs, meaning that everyone is at war with someone. It's still good to use AI_isWarPossible in AI code instead of checking the "Always Peace" option, so I didn't revert that part.