More bugs in deal cancellation; and amend previous commit.
The way that CvPlayerAI::AI_bonusTrade currently works, iChange needs to be set to 0 when considering cancellation. The previous commit didn't get that right.

Fixed an issue (introduced by [advc.074]) with resources not being shown on the trade table when the AI announces that it has canceled a deal.
See the last paragraph of this post for more info:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-12#post-15492585
The fix was somewhat complicated to implement because the call to CvPlayer::buildTradeTable comes from the EXE.

Added some wrapper functions to CvDeal to make it easier to iterate over trade items. [advc.003] Inlined some of the old (exported) functions in the process. [advc.003f] The new interface is, so far, only used for the [advc.074] bugfix, but there are (many?) other places in the code where it should be used to remove duplicate code – to be done (perhaps better wait with that until after v0.96).

Fixed a recently introduced bug in [advc.072] that had removed the turns-to-cancel string from the hover text under "what are the current deals we have" (trade table). The recent change had broken CvDeal::isCancelable.