Changed the order of culture percentages in tile help text
The active player's culture is now always listed first (if tile owner – no change) or second, even if the active player is not in slot 0 (e.g. in a scenario). The other players (non-active, not tile owner) are ordered by tile culture (instead of slot id). Wrote code for putting all percentages on a single line but didn't like the result; left it commented out. [advc.099g]

Added a comment about an open issue with the previous commit and made a small amendment that doesn't really hurt or help. [advc.074]