Update attitude cache after Auto Play; and a minor bugfix.
Previously, the AI attitude cache was only updated at the start of AI Auto Play. [advc.127]

Bug in [advc.034] ("disengagement" agreements): Had wasted a bit of computing time when that feature was disabled through XML (which it is by default).

And some comments added (about the performance of individual functions and about the BULL airport icon).