Core functionality; UI support still work in progress.
The defender randomization module makes all but 4 units in each tile unavailable as defenders. The available units are chosen at random with a strong bias for healthy units. When a tile is attacked, the best available unit defends (Lead From Behind algorithm as before).

Implementation based on the concept I had sketched here:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-11#post-15480906
which, in turn, is based on devolution's prototype:
https://github.com/ErikSandbraaten/AdvCiv/tree/limited_attacks_and_defender_set

I think my design of the DefenderSelector class and its interaction with CvPlot are sound. The interaction with CvGameTextMgr may need work (but not much).

Hardly tested. The defender set seems to get computed correctly. Performance hasn't been measured. The cache is currently recomputed on every access, but that can easily be changed once cache update calls are added (see to-do list in DefenderSelector::cacheDefenders).

UI support is still poor. Unavailable units are shown in gray – that's all. Won't work correctly for very large stacks when multiple units are shown in one line of hover text.
Tbd.: The available units need to be shown on top. Not sure if civilians and third-party units should be grayed out; currently they aren't (on purpose). Need to think the various stack compositions through and then rewrite some of the code, especially in CvGameTextMgr::setPlotListHelpPerOwner. Some combat messages should also be added.

I haven't even thought about the AI yet.