Worked out the open issues with help text
The available units are now listed on top and don't get aggregated or collapsed when there is a high number of units to list.

Did a profiler run; negligible computational effort. The cache seems unnecessary then. So I'm not going to add any DefenderSelector::update calls for the time being.

To be done (but I won't do most of it soon):
* Naval stacks/ submarines are still mostly untested, but the logic for handling them is in place. Also to be tested: damage limits, air strikes, missiles.
* Help text in Debug mode may still have some minor issues.
* Combat messages (on-screen and combat log)
* AI: I guess any stack-vs-stack heuristics will have to favor the attacking side more.
* Configuration options (on/off switch, number of defenders, gray-out color)
* Documentation