Made most CvGlobals member functions const
"GC" (global context) is now defined as a CvGlobals const& – was CvGlobals&. For calling non-const functions, CvGlobals::getInstance() needs to be used. That should serve as a warning, as most classes have no business changing global data. Not sure if const-ness can also improve performance in this case; certainly can't hurt.

I haven't added const qualifiers to all CvGlobals functions that could have them, but most. Had to overload a lot of DllExports. The functions that provide references to singleton objects not owned by CvGlobals (in particular getGame, getInitCore and the getters for the various pathfinder instances) still return non-const references.

The non-synchronized PRNG, which is owned by CvGlobals, is now accessed through a truly global function. Meaning that GC.getAsyncRand() calls throughout the code (not many really; mainly in CvPlayerAI.cpp) have been replaced with just getAsyncRand().

The implementation files that set most of the data stored at CvGlobals – CvInfos.cpp and CvXMLLoadUtilitySet.cpp – simply redefine "GC" as a non-const reference, so nothing else has to be changed in those two files.

Moved a little bit of code from CvMap to CvGlobals to remove a dependency of CvMap on CvLandscapeInfo. In preparation of splitting up CvInfos.h. [advc.003x]