v0.96c: Bug fixed AI resource evaluation
This bug had gone unnoticed by me for a long time because it had only occurred in final release builds. Apparently, the arch:SSE2 compiler option can lead to negative results from std::pow<double> when the base is negative. In any case, a negative base is a coding error. It seems that the main effect of the error was erratic cancellation of resource trades by the AI.

Updated DLL included.